backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
face_recognition
fastapi==0.135.3
fastapi-cli==0.0.24
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pyotp==2.9.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.4.0
rich==14.3.3
sentry-sdk==2.57.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
